Add prepare-commit-msg git hook
Before git v2.3, the "Conflicts" section that was part of a merge commit
message (if the merge generated any conflicts) was uncommented out,
which was really helpful to see if any merges resulted in conflicts.

Since git v2.3, this "Conflicts" section was commented out by default
(wtf why?!). So add this hook to edit the commit-msg to ensure the
Conflicts block is retained.