Fix :Flake command for Vim 7.4.1013+
Vim 7.4.1013 onwards tries to use the `'errorformat'` setting if it is
defined. However, this issue makes `:lexpr` and `:cexpr` stop populating
the location/quickfix list if a different command was used which the
errorformat was not designed for.

Thus, this small hack, first saves the errorformat, clears it, and then
re-applies it back again after executing `:lexpr`.

Source: https://github.com/vim/vim/issues/569